% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadixSort.R
\name{RadixSort}
\alias{RadixSort}
\title{Tri par Radix Sort}
\usage{
RadixSort(arr)
}
\arguments{
\item{arr}{Le vecteur d'entiers à trier.}
}
\value{
Le vecteur d'entiers trié.
}
\description{
Cette fonction trie un vecteur d'entiers en utilisant l'algorithme Radix Sort.
}
\details{
L'algorithme Radix Sort est un algorithme de tri non comparatif
qui trie les éléments en examinant les chiffres de leurs représentations
décimales.
}
\examples{
arr <- c(170, 45, 75, 90, 802, 24, 2, 66)
RadixSort(arr)
}
