% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_RadixSort}
\alias{rcpp_RadixSort}
\title{Fonction de tri Radix Sort (implémenté en C++)}
\usage{
rcpp_RadixSort(arr)
}
\arguments{
\item{arr}{Le vecteur d'entiers à trier.}
}
\value{
Le vecteur d'entiers trié.
}
\description{
Cette fonction trie un vecteur d'entiers en utilisant l'algorithme Radix Sort.
}
\examples{
arr <- c(170, 45, 75, 90, 802, 24, 2, 66)
rcpp_RadixSort(arr)
}
