% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_TriFusion}
\alias{rcpp_TriFusion}
\title{Tri par fusion (implémenté en C++)}
\usage{
rcpp_TriFusion(tableau)
}
\arguments{
\item{tableau}{Le vecteur d'entiers à trier.}
}
\value{
Le vecteur d'entiers trié.
}
\description{
Cette fonction trie un vecteur d'entiers en utilisant l'algorithme de tri fusion.
}
\details{
L'algorithme de tri fusion est un algorithme de tri efficace
et stable qui divise le vecteur en deux moitiés, trie récursivement
chaque moitié, puis fusionne les deux moitiés triées pour produire
le résultat final trié.
}
\examples{
tableau <- c(38, 27, 43, 3, 9, 82, 10)
rcpp_TriFusion(tableau)
}
